型定義BoardPropsをBoard/index.tsxに移動
