+++ types/index.ts
@@ -1,12 +1,3 @@
 export type Mark = 'X' | 'O' | null;
 
 export type Coordinates = { x: number; y: number } | null;
-
-export type BoardProps = {
-  xIsNext: boolean;
-  squares: Mark[];
-  lines: number[][];
-  boardSize: number;
-  isDraw: boolean;
-  handlePlay: (nextSquares: Mark[], coordinates: Coordinates) => void;
-};
